stmodel/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
numpy>=1.24.0
psycopg2-binary>=2.9.0